requirements.txt
jupyter
pandas
fitdecode
gpxpy
numpy
matplotlib
dotenv
mysql.connector